adding mobile, gc pages

+++ src/components/SelectedRiders.vue
@@ -1,15 +1,21 @@
 <template>
   <div class="selectedInfo">
-    <div>Geselecteerd: {{ riderStore.ridersCount }} van 8 renners</div>
+    <div :class="riderStore.ridersCount > 8 ? 'red' : ''">
+      Geselecteerd: {{ riderStore.ridersCount }} van 8 renners
+    </div>
+
     <div class="btn btn-action" @click="sendSelection">{{ sendButton }}</div>
-    <div class="btn btn-danger">Wis selectie</div>
+    <div class="btn btn-danger" @click="wisSelection">Wis selectie</div>
+    <div v-show="errorMsg.length >= 1" class="errormessage">
+      {{ errorMsg }}
+    </div>
   </div>
   <div class="selectedCyclists">
     <RennerSmallCard
       v-for="(renner, index) in riderStore.riders.renners"
       :renner="renner"
       :icon="'minus'"
-      @click="deleteRenner(index)"
+      @click="deleteRenner(index, renner)"
       class="selectie"
     />
   </div>
@@ -23,6 +29,7 @@ import axios from "axios";
 import RennerSmallCard from "../components/RennerSmallCard.vue";
 
 import { useCyclistStore } from "../stores/selectedRiders";
+import { useStartlistStore } from "../stores/startlistRace";
 import { useAuthStore } from "../stores/userAuth";
 import router from "../router";
 
@@ -31,15 +38,27 @@ export default defineComponent({
 
   setup() {
     const riderStore = useCyclistStore();
+    const startlistStore = useStartlistStore();
     const stage_id = riderStore.riders.stage;
     const route = useRoute();
     const raceInfo = ref();
     const avoidRouteLeave = ref(false);
-    const user_id = ref(useAuthStore().getUserID);
+    // const user_id = ref(useAuthStore().getUserID);
+    const user_id = ref(1);
+    const errorMsg = ref("");
 
     const sendButton = ref("Verstuur selectie");
 
     function sendSelection() {
+      if (riderStore.ridersCount > 8) {
+        errorMsg.value =
+          "Je hebt teveel renners geselecteerd! Kies er maximaal 8.";
+        setTimeout(() => {
+          errorMsg.value = "";
+        }, 5000);
+        return;
+      }
+
       sendButton.value = "Versturen...";
 
       axios.delete(
@@ -75,6 +94,30 @@ export default defineComponent({
       });
     }
 
+    function wisSelection() {
+      riderStore.riders.renners.forEach((renner) => {
+        startlistStore.removeSelected(renner);
+        riderStore.removeSelected(renner);
+      });
+      riderStore.madeChange();
+    }
+
+    function deleteRenner(rennerIndex, renner) {
+      riderStore.madeChange();
+      startlistStore.removeSelected(renner);
+      riderStore.removeRiderByIndex(rennerIndex);
+
+      return;
+    }
+
+    function setCyclistSelected(cyclist) {
+      const selectedRider = startlistStore.teams[cyclist.team_name].find(
+        ({ cyclist_id }) => cyclist.cyclist_id == cyclist_id
+      );
+
+      selectedRider.selected = true;
+    }
+
     onMounted(async () => {
       try {
         const { data: entryData } = await axios({
@@ -93,6 +136,7 @@ export default defineComponent({
             const newRenner = { ...renner, selected: true };
 
             riderStore.addRider(newRenner);
+            setCyclistSelected(newRenner);
           });
         }
         if (raceData) {
@@ -103,7 +147,7 @@ export default defineComponent({
       }
     });
     onBeforeRouteLeave((to, from, next) => {
-      if (avoidRouteLeave.value || riderStore.ridersCount == 0) {
+      if (avoidRouteLeave.value || riderStore.selectionChanges == false) {
         next();
       } else {
         const answer = window.confirm(
@@ -120,7 +164,17 @@ export default defineComponent({
         }
       }
     });
-    return { avoidRouteLeave, raceInfo, riderStore, sendSelection, sendButton };
+    return {
+      avoidRouteLeave,
+      deleteRenner,
+      wisSelection,
+      raceInfo,
+      riderStore,
+      startlistStore,
+      sendSelection,
+      sendButton,
+      errorMsg,
+    };
   },
 });
 </script>
@@ -128,5 +182,10 @@ export default defineComponent({
 <style lang="scss">
 .selectedCyclists {
   margin-top: 2rem;
+}
+
+.red {
+  color: var(--clr-accent);
+  font-weight: 900;
 }
 </style>
